refactor: simplify chord interface
