--- frontend/src/app/interfaces/chord-data.ts
@@ -1,7 +1,0 @@
-import { ChordNode } from '../interfaces/chord-node';
-import { ChordLink } from './chord-link';
-
-export interface ChordData {
-  nodes: ChordNode[];
-  links: ChordLink[];
-}
