--- frontend/src/app/interfaces/chord-link.ts
@@ -1,4 +1,0 @@
-export interface ChordLink {
-  source: string;
-  target: string;
-}
