--- frontend/src/app/interfaces/chord-node.ts
@@ -1,5 +1,0 @@
-export interface ChordNode {
-  name: string;
-  group: string;
-  id?: string;
-}
