+++ frontend/src/app/interfaces/chord.ts
@@ -1,0 +1,15 @@
+export interface ChordNode {
+  readonly name: string;
+  readonly group: string;
+  readonly id?: string;
+}
+
+export interface ChordLink {
+  readonly source: string;
+  readonly target: string;
+}
+
+export interface ChordData {
+  readonly nodes: ChordNode[];
+  readonly links: ChordLink[];
+}
